Add error type for required secrets
