+++ azcfg.go
@@ -47,7 +47,7 @@ func parse(d any, client SecretsClient) error {
 
 	if err := setFields(v, secrets); err != nil {
 		if len(required) > 0 {
-			return errors.New(requiredErrorMessage(secrets, required))
+			return newRequiredError(secrets, required)
 		}
 		return err
 	}
@@ -207,37 +207,4 @@ func isRequired(values []string) bool {
 		return false
 	}
 	return values[1] == required
-}
-
-// requiredErrorMessage builds a message based on the provided []string.
-func requiredErrorMessage(secrets map[string]string, required []string) string {
-	if len(required) == 0 {
-		return ""
-	}
-
-	req := make([]string, 0)
-	for _, r := range required {
-		if len(secrets[r]) == 0 {
-			req = append(req, r)
-		}
-	}
-
-	var message strings.Builder
-	l := len(req)
-	if l == 1 {
-		message.WriteString("secret: " + req[0] + " is required")
-		return message.String()
-	}
-	message.WriteString("secrets: ")
-	for i, r := range req {
-		message.WriteString(r)
-		if i < l-1 && l > 2 && i != l-2 {
-			message.WriteString(", ")
-		}
-		if i == l-2 {
-			message.WriteString(" and ")
-		}
-	}
-	message.WriteString(" are required")
-	return message.String()
 }
